User commit

+++ src/main/java/com/example/demo/service/impl/ValidationServiceimpl.java
@@ -17,7 +17,7 @@ public class ValidationServiceimpl implements ValidationService{
     }
 
     @Override
-    public ValidationEntity getData(int id){
+    public ValidationEntity getData(Long id){
         return student.findById(id).orElseThrow(()-> new ValidationException("Invalid Id" + id));
